Cleanup unreferenced rating plans and rating profiles
cgr-loader (and load_tp_from_stordb api method) do not delete keys from datadb
(redis), they just update keys with new values.

While this is not a problem for destinations (apart from datadb with stale data),
it is a problem for rating plans and rating profiles: changes in UI do not apply
as keys persist in datadb.

This patch does:

- New API method: clean_datadb_from_unused_tp.

- New flag for cgr-loader: -cleanup.

- New flag for load_tp_from_stordb API method: -cleanup.

All of them remove unreferenced rating plans and profiles from datadb.

(Note: only valid for IvozProvider as long as tpid is unique globally)